Create validation mechanism and update operations
The validation data can now be accessed by creating another queue with another directory of data. Then a boolean operator is passed into the feed_dict which signals which queue to pull the data from.

Two update operators were added. One for updating the forward cells in a blstm and one for the backward cells in a blstm.